requirements.txt
fastapi
uvicorn
tensorflow
numpy
Pillow
python-multipart